+++ src/Http/HttpServiceProvider.php
@@ -5,6 +5,7 @@ namespace BristolSU\Support\Http;
 use BristolSU\Support\Http\View\InjectJavascriptVariables;
 use BristolSU\Support\Http\View\InjectOldInput;
 use BristolSU\Support\Http\View\InjectValidationErrors;
+use BristolSU\Support\Http\View\Router\InjectNamedRoutes;
 use Illuminate\Support\Facades\View;
 use Illuminate\Support\ServiceProvider;
 
@@ -24,6 +25,7 @@ class HttpServiceProvider extends ServiceProvider
         View::composer(['bristolsu::base'], InjectJavascriptVariables::class);
         View::composer(['bristolsu::base'], InjectValidationErrors::class);
         View::composer(['bristolsu::base'], InjectOldInput::class);
+        View::composer(['bristolsu::base'], InjectNamedRoutes::class);
     }
 
 }