+++ src/Http/View/Router/NamedRouteRetrieverCache.php
@@ -1,0 +1,32 @@
+<?php
+
+namespace BristolSU\Support\Http\View\Router;
+
+use Illuminate\Contracts\Cache\Repository;
+
+class NamedRouteRetrieverCache implements NamedRouteRetrieverInterface
+{
+
+    /**
+     * @var NamedRouteRetrieverInterface
+     */
+    private NamedRouteRetrieverInterface $retriever;
+    /**
+     * @var Repository
+     */
+    private Repository $cache;
+
+    public function __construct(NamedRouteRetrieverInterface $retriever, Repository $cache)
+    {
+        $this->retriever = $retriever;
+        $this->cache = $cache;
+    }
+
+    public function all(): array
+    {
+        return $this->cache->remember(NamedRouteRetrieverCache::class, 600, function() {
+            return $this->retriever->all();
+        });
+    }
+
+}