+++ src/Http/View/Router/NamedRouteRetrieverInterface.php
@@ -1,0 +1,11 @@
+<?php
+
+
+namespace BristolSU\Support\Http\View\Router;
+
+
+interface NamedRouteRetrieverInterface
+{
+    public function all(): array;
+
+}